refactor(frontend): split agent orchestrator operations
