+++ packages/frontend/src/state/operations/agent-orchestrator/lifecycle/load-sessions.ts
@@ -1,4 +1,5 @@
 import type { RepoPromptOverrides, TaskCard } from "@openducktor/contracts";
+import type { QueryClient } from "@tanstack/react-query";
 import type { Dispatch, MutableRefObject, SetStateAction } from "react";
 import { appQueryClient } from "@/lib/query-client";
 import { loadAgentSessionListFromQuery } from "@/state/queries/agent-sessions";
@@ -32,6 +33,7 @@ type CreateLoadAgentSessionsArgs = {
   loadRepoPromptOverrides: (workspaceId: string) => Promise<RepoPromptOverrides>;
   loadTaskDocuments?: (repoPath: string, taskId: string) => Promise<TaskDocuments>;
   agentSessionPresenceStore?: AgentSessionPresenceStore;
+  queryClient?: QueryClient;
 };
 
 export const createLoadAgentSessions = ({
@@ -48,6 +50,7 @@ export const createLoadAgentSessions = ({
   loadRepoPromptOverrides,
   loadTaskDocuments: _loadTaskDocuments,
   agentSessionPresenceStore,
+  queryClient = appQueryClient,
 }: CreateLoadAgentSessionsArgs): ((
   taskId: string,
   options?: AgentSessionLoadOptions,
@@ -176,7 +179,7 @@ export const createLoadAgentSessions = ({
           loadPersistedRecords: async () =>
             options?.persistedRecords
               ? Promise.resolve(options.persistedRecords)
-              : loadAgentSessionListFromQuery(appQueryClient, repoPath, taskId),
+              : loadAgentSessionListFromQuery(queryClient, repoPath, taskId),
           loadRepoPromptOverrides,
         });
       if (isStaleRepoOperation()) {
